Added "volatile" qualifier to the variable buzzerSequence.
In general, all variables that are written in an ISR and read in the main loop
need to be marked as volatile.  This commit fixes a particular bug we
encountered where GCC (if LTO and other optimizations are turned on) would
optimize loops like this:

    while (!buzzer.isPlaying());

GCC moved the loading of buzzerSequence to be outside of the loop in order to
make the loop faster, and that caused the program to behave incorrectly.